cleaner EOF check
